requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.3
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Authlib==1.6.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.4.1
cli_exit_tools==1.2.7
click==8.1.7
comm==0.2.0
contourpy==1.2.0
courlan==1.3.2
cryptography==45.0.5
cycler==0.12.1
cyclopts==3.22.2
dataclasses-json==0.6.7
dateparser==1.2.0
dateutils==0.6.12
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.0.1
fake-winreg==1.6.3
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastapi-responses==0.2.1
fastjsonschema==2.19.0
fastmcp==2.10.5
filelock==3.18.0
fireworks-ai==0.15.11
fonttools==4.47.0
fqdn==1.5.1
frozendict==2.3.10
frozenlist==1.5.0
fsspec==2025.7.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.25.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
h11==0.14.0
hf-xet==1.1.5
html2text==2024.2.26
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
huggingface-hub==0.34.3
idna==3.6
iniconfig==2.1.0
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.5.1
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jusText==3.0.1
kiwisolver==1.4.5
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-community==0.3.14
langchain-core==0.3.63
langchain-fireworks==0.2.6
langchain-huggingface==0.2.0
langchain-openai==0.3.19
langchain-pinecone==0.2.11
langchain-tests==0.3.20
langchain-text-splitters==0.3.5
langsmith==0.2.11
lib-detect-testenv==2.0.8
lib-platform==1.2.10
lib-registry==2.0.10
lxml==5.3.0
lxml_html_clean==0.4.1
MainContentExtractor==0.0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.24.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mcp==1.11.0
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.5
notebook==7.0.6
notebook_shim==0.2.3
nsepy==0.8
nsetools==1.0.11
numpy==1.26.4
openai==1.99.1
openapi-pydantic==0.5.1
orjson==3.10.13
overrides==7.4.0
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
peewee==3.17.0
pexpect==4.9.0
Pillow==10.1.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
platformdirs==4.1.0
playwright==1.49.1
pluggy==1.6.0
posthog==3.7.5
prometheus-client==0.19.0
prompt-toolkit==3.0.43
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==5.9.7
psycopg==3.2.10
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.11.7
pydantic-settings==2.7.1
pydantic_core==2.33.2
pyee==12.0.0
Pygments==2.17.2
pyparsing==3.1.1
pypdf==5.9.0
pyperclip==1.9.0
pytest==8.4.1
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==4.0.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-rst==1.3.1
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.15.2
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
Send2Trash==1.8.2
sentence-transformers==5.0.0
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.36
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.47.1
sympy==1.14.0
syrupy==4.9.1
ta==0.11.0
tenacity==8.5.0
terminado==0.18.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tinycss2==1.2.1
tld==0.13
tokenizers==0.21.4
torch==2.7.1
tornado==6.4
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.0
transformers==4.55.0
typer==0.16.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.1.0
uvicorn==0.35.0
uvloop==0.21.0
vcrpy==7.0.0
watchfiles==1.1.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
twilio==9.8.1
sounddevice==0.5.2
